+++ notify-stack/src/main/java/com/dwarfeng/notify/stack/dao/UserDao.java
@@ -4,6 +4,7 @@ import com.dwarfeng.notify.stack.bean.entity.User;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
+import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
 
 /**
  * 用户数据访问层。
@@ -11,5 +12,5 @@ import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
  * @author DwArFeng
  * @since 1.0.0
  */
-public interface UserDao extends BatchBaseDao<StringIdKey, User>, EntireLookupDao<User> {
+public interface UserDao extends BatchBaseDao<StringIdKey, User>, EntireLookupDao<User>, PresetLookupDao<User> {
 }